server/requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
Flask-Migrate
flask-admin
 